fix(web): #1226 — Vercel preview build without Clerk key no longer crashes
Preview builds have no NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY set. Root layout
skips <ClerkProvider>, but several components called useAuth()
unconditionally and crashed at prerender.

Three fixes, one root-cause pattern (env-gate useAuth like WorkspaceProvider
already does):

- components/marketing/CtaLink.tsx: split into SignUpLink / CtaLinkWithAuth
  so useAuth is only called when Clerk is wired. MarketingNav renders CtaLink
  on every marketing page, so this alone unblocked every static marketing
  route.
- (marketing)/playbooks/page.tsx: gate useAuth via a compile-time constant.
  Only Clerk-using marketing page; same pattern.
- (app)/layout.tsx: split into a thin server layout that exports
  dynamic = "force-dynamic" and a client shell (AppSegment.tsx). Route
  Segment Config exports are ignored on client components in Next.js 15,
  so the split is mandatory to skip prerender for the auth-required subtree.

Verified: NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY= npx next build succeeds and
generates 61/61 static pages. Production build (with the key set) unaffected.
