+++ apps/web/src/components/marketing/CtaLink.tsx
@@ -1,7 +1,23 @@
 "use client"
 import { useAuth } from "@clerk/nextjs"
 
-export function CtaLink({ className, children }: { className: string; children?: React.ReactNode }) {
+// ponytail: split so useAuth is only called when Clerk is actually wired up.
+// Preview builds without NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY get the plain link;
+// prod/preview with Clerk get the auth-aware version. Same pattern as
+// WorkspaceProvider (lib/WorkspaceContext.tsx).
+const CLERK_ENABLED = !!process.env.NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY
+
+type Props = { className: string; children?: React.ReactNode }
+
+export function CtaLink(props: Props) {
+  return CLERK_ENABLED ? <CtaLinkWithAuth {...props} /> : <SignUpLink {...props} />
+}
+
+function SignUpLink({ className, children }: Props) {
+  return <a href="/sign-up" className={className}>{children ?? "Start Free"}</a>
+}
+
+function CtaLinkWithAuth({ className, children }: Props) {
   const { isSignedIn, isLoaded } = useAuth()
   if (!isLoaded) return <span className={className} aria-hidden />
   return isSignedIn
